apps/backend/admin-api/requirements.txt
fastapi>=0.110.0
fastapi-admin>=1.0.4
uvicorn[standard]>=0.29.0
tortoise-orm>=0.20.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 